moved videoCardSchema for faster hot reloads
